+++ db/migrate/20130101000001_create_users.rb
@@ -35,9 +35,6 @@ class CreateUsers < ActiveRecord::Migration[5.2]
       t.string :current_sign_in_ip
       t.string :last_sign_in_ip
 
-      ## Encryptable
-      t.string :password_salt
-
       ## Confirmable
       t.string   :confirmation_token
       t.datetime :confirmed_at