+++ lib/generators/ab_admin/install/templates/models/user.rb
@@ -1,7 +1,6 @@
 class User < ApplicationRecord
   devise :database_authenticatable, :confirmable, :lockable,
-         :recoverable, :rememberable, :trackable, :validatable, :registerable,
-         :encryptable, encryptor: :sha512
+         :recoverable, :rememberable, :trackable, :validatable, :registerable
 
   include AbAdmin::Concerns::AdminAddition
   include AbAdmin::Models::User