+++ spec/dummy/db/schema.rb
@@ -264,7 +264,6 @@ ActiveRecord::Schema.define(version: 2019_10_29_085351) do
     t.datetime "last_sign_in_at"
     t.string "current_sign_in_ip"
     t.string "last_sign_in_ip"
-    t.string "password_salt"
     t.string "confirmation_token"
     t.datetime "confirmed_at"
     t.datetime "confirmation_sent_at"